add spectral reg + package setup
