+++ setup.py
@@ -1,0 +1,5 @@
+from setuptools import setup, find_packages
+setup(
+    name = 'memorization',
+    packages = find_packages(),
+)
